Refactor pixel processing to use stripe-based approach
Replace full-image IntBuffer allocation with stripe-based processing
to dramatically reduce peak Java heap usage during bitmap comparison.

- Process bitmaps in horizontal stripes using Bitmap.createBitmap()
  sub-regions and copyPixelsToBuffer(), keeping per-stripe heap usage
  under ~1 MB instead of ~19 MB for full-image buffers
- Stripe height is dynamically calculated based on available heap memory
- Revert ByteBuffer.allocateDirect() change (no benefit on Android where
  direct buffers use VMRuntime.newNonMovableArray on the Java heap)
- Add mockBitmapCreateBitmap() test helper and update all processor tests

+++ Library/src/main/java/dev/testify/core/exception/LowMemoryException.kt
@@ -50,7 +50,7 @@ import dev.testify.core.formatDeviceString
 @Suppress("ktlint:standard:argument-list-wrapping")
 class LowMemoryException(
     targetContext: Context,
-    requestedAllocation: Long,
+    requestedAllocation: Int,
     memoryInfo: String,
     cause: OutOfMemoryError
 ) : TestifyException(
